Change junit.jupiter.execution.parallel.mode.default = "concurrent". Speed up, parellilize tests where possible.
